Struct turborepo_lib::Args

pub struct Args {

    pub version: bool,
    pub skip_infer: bool,
    pub api: Option<String>,
    pub color: bool,
    pub cpu_profile: Option<String>,
    pub cwd: Option<PathBuf>,
    pub heap: Option<String>,
    pub login: Option<String>,
    pub no_color: bool,
    pub preflight: bool,
    pub team: Option<String>,
    pub token: Option<String>,
    pub trace: Option<String>,
    pub verbosity: Verbosity,
    pub check_for_update: bool,
    pub test_run: bool,
    pub run_args: Option<RunArgs>,
    pub command: Option<Command>,

}

Fields
version: boolskip_infer: bool

Skip any attempts to infer which version of Turbo the project is configured to use

api: Option<String>

Override the endpoint for API calls

color: bool

Force color usage in the terminal

cpu_profile: Option<String>

Specify a file to save a cpu profile

cwd: Option<PathBuf>

The directory in which to run turbo

heap: Option<String>

Specify a file to save a pprof heap profile

login: Option<String>

Override the login endpoint

no_color: bool

Suppress color usage in the terminal

preflight: bool

When enabled, turbo will precede HTTP requests with an OPTIONS request for authorization

team: Option<String>

Set the team slug for API calls

token: Option<String>

Set the auth token for API calls

trace: Option<String>

Specify a file to save a pprof trace

verbosity: Verbosity

verbosity

check_for_update: bool

Force a check for a new version of turbo

test_run: boolrun_args: Option<RunArgs>command: Option<Command>
Implementations

impl Args

pub fn new() -> Result<Self>

impl Args

pub fn ui(&self) -> UI